Moved NQL string escaping into a shared package
ref https://linear.app/ghost/issue/ONC-1920

There were three copies of this helper - in admin-x-framework, the Ember
admin and ghost/core - and they had started to drift, with two escaping
only single quotes and the third escaping both.

The two rules are easy to get wrong in opposite directions, which is why
one implementation is safer than three:

- NQL rejects a bare `"` inside a single-quoted string, so escaping only
  `'` left any value containing a double quote unparseable - that affects
  every typeahead search box in admin, not just link editing
- backslashes must not be escaped, because NQL has no `\\` escape and
  doubling one would query a different value

Unifying on both-quote escaping cannot regress a working case: a value
containing `"` today either fails to parse or silently matches the wrong
thing, and no call site pre-escapes.

The package is ESM-only and resolved through the `source` condition in dev
and test, matching custom-field-types. ghost/core reaches it from
CommonJS, which Node handles on the pinned ^22.23.1 engine - the same way
admin-api-schema is already consumed.
