+++ apps/admin-x-framework/package.json
@@ -110,6 +110,7 @@
         "@tanstack/react-query": "catalog:",
         "@tinybirdco/charts": "0.3.0",
         "@tryghost/custom-field-types": "workspace:*",
+        "@tryghost/nql-string": "workspace:*",
         "@tryghost/shade": "workspace:*",
         "bson-objectid": "catalog:",
         "react": "catalog:",
