+++ apps/ember-admin/app/utils/escape-nql-string.js
@@ -1,10 +1,4 @@
-// Escape a search term and wrap it in single quotes for safe embedding in an
-// NQL filter, e.g. `title:~<result>`.
-//
-// Only single quotes are escaped: the NQL lexer treats just `\'`/`\"` as
-// escapes and reads a lone backslash literally. Escaping every quote prevents
-// breakout, and backslashes must not be doubled because that corrupts terms
-// containing a backslash.
-export function escapeNqlString(term) {
-    return '\'' + String(term).split('\'').join('\\\'') + '\'';
-}
+// Re-exported so existing `ghost-admin/utils/escape-nql-string` imports keep
+// working. The implementation lives in @tryghost/nql-string so that the admin,
+// the Ember admin and ghost/core all escape filter values identically.
+export {escapeNqlString} from '@tryghost/nql-string';
