+++ apps/ember-admin/package.json
@@ -55,6 +55,7 @@
         "@tryghost/koenig-lexical": "workspace:*",
         "@tryghost/limit-service": "catalog:",
         "@tryghost/nql": "catalog:",
+        "@tryghost/nql-string": "workspace:*",
         "@tryghost/string": "catalog:",
         "@tryghost/timezone-data": "catalog:",
         "animejs": "3.2.2",
