+++ ghost/core/core/server/services/link-tracking/link-click-tracking-service.js
@@ -4,7 +4,7 @@ const PostLink = require('./post-link');
 const ObjectID = require('bson-objectid').default;
 const errors = require('@tryghost/errors');
 const nql = require('@tryghost/nql');
-const {escapeNqlString} = require('../../lib/nql-string');
+const {escapeNqlString} = require('@tryghost/nql-string');
 const _ = require('lodash');
 const tpl = require('@tryghost/tpl');
 const moment = require('moment');
