+++ ghost/core/package.json
@@ -124,6 +124,7 @@
         "@tryghost/nodemailer": "2.3.1",
         "@tryghost/nql": "catalog:",
         "@tryghost/nql-lang": "catalog:",
+        "@tryghost/nql-string": "workspace:*",
         "@tryghost/parse-email-address": "workspace:*",
         "@tryghost/pretty-cli": "3.3.2",
         "@tryghost/prometheus-metrics": "1.0.8",
