+++ packages/nql-string/eslint.config.mjs
@@ -1,0 +1,3 @@
+import {nodeLibConfig} from '@internal/cfg-eslint';
+
+export default nodeLibConfig();
