+++ packages/nql-string/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@internal/cfg-typescript/esm.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "build"
+  },
+  "include": ["src/**/*"]
+}
