+++ packages/nql-string/vitest.config.ts
@@ -1,0 +1,3 @@
+import {createVitestConfig} from '@internal/cfg-vitest';
+
+export default createVitestConfig();
